Archive/requirements/base.txt
django==5.1.4
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-elasticsearch-dsl==8.1
django-filter==24.3
django-jazzmin==3.0.1
django-redis==5.4.0
django-rosetta==0.10.1
djangorestframework==3.15.2
drf-yasg==1.21.8
elasticsearch==8.9.0
elasticsearch-dsl==8.9.0
flower==2.0.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-environ
python-telegram-bot==21.9
redis==5.2.1
tika==2.6.0
whitenoise
Pillow
pdf2image==1.17.0
psutil==5.9.8
